Move take_non_fungible method into NonFungibleVault blueprint
