+++ radix-engine/src/blueprints/resource/non_fungible/non_fungible_vault.rs
@@ -50,6 +50,39 @@ impl NonFungibleVaultBlueprint {
         Ok(Bucket(bucket_id))
     }
 
+    pub fn take_non_fungibles<Y>(
+        receiver: &RENodeId,
+        non_fungible_local_ids: BTreeSet<NonFungibleLocalId>,
+        api: &mut Y,
+    ) -> Result<Bucket, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let info = VaultInfoSubstate::of(receiver, api)?;
+
+        // Take
+        let taken =
+            NonFungibleVault::take_non_fungibles(receiver, &non_fungible_local_ids, api)?;
+
+        // Create node
+        let bucket_id = api.new_object(
+            BUCKET_BLUEPRINT,
+            vec![
+                scrypto_encode(&BucketInfoSubstate {
+                    resource_address: info.resource_address,
+                    resource_type: info.resource_type,
+                })
+                    .unwrap(),
+                scrypto_encode(&LiquidFungibleResource::default()).unwrap(),
+                scrypto_encode(&LockedFungibleResource::default()).unwrap(),
+                scrypto_encode(&taken).unwrap(),
+                scrypto_encode(&LockedNonFungibleResource::default()).unwrap(),
+            ],
+        )?;
+
+        Ok(Bucket(bucket_id))
+    }
+
     pub fn put<Y>(receiver: &RENodeId, bucket: Bucket, api: &mut Y) -> Result<(), RuntimeError>
     where
         Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
