+++ radix-engine/src/blueprints/resource/resource_manager_util.rs
@@ -146,7 +146,7 @@ fn build_access_rules(
         DenyAll,
     );
     vault_access_rules.set_group_and_mutability(
-        MethodKey::new(NodeModuleId::SELF, VAULT_TAKE_NON_FUNGIBLES_IDENT),
+        MethodKey::new(NodeModuleId::SELF, NON_FUNGIBLE_VAULT_TAKE_NON_FUNGIBLES_IDENT),
         "withdraw",
         DenyAll,
     );
