+++ scrypto/src/resource/vault.rs
@@ -123,8 +123,8 @@ impl ScryptoVault for Vault {
         let rtn = env
             .call_method(
                 &RENodeId::Object(self.0),
-                VAULT_TAKE_NON_FUNGIBLES_IDENT,
-                scrypto_encode(&VaultTakeNonFungiblesInput {
+                NON_FUNGIBLE_VAULT_TAKE_NON_FUNGIBLES_IDENT,
+                scrypto_encode(&NonFungibleVaultTakeNonFungiblesInput {
                     non_fungible_local_ids: non_fungible_local_ids.clone(),
                 })
                 .unwrap(),
